requirements.txt
requests==2.26.0

# not psycopg2, just binary
psycopg2-binary==2.9.3
sqlalchemy